+++ changelog.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Updated Briggs support to include text output for better MQTT support
 - Corrected on bug in fuel calculation for 32kW Evolution based Propane Liquid Cooled units
 - Corrected bug in Briggs and Stratton controller definition. Changed name of some gauges to better describe the value.
+- Added new alarm code for Nexus Evolution
 
 ## V1.19.03 -2024-04-19
 - New add on program genmqttin.py. See wiki and add on page for details